Move all ReversibleIBookAction classes into new package
